Added multi-screen capabilities
Why:
So we can switch between different screens

How:
Created a system of multiple screens being overlayed over each other.
Every time a button on the sidebar is pressed, a new frame is drawn on
top of the current screen frame. This allows not only for changing
screens, but also preserving what was previously on the screen.

Tags:[feat][refactor]