Added searching for types other than items
Why:
So the program can search for monsters, vehicle parts, mutations,
bionics, and martial arts.

How:
Rather than keeping all the JSON in one big dictionary, there is now a
dictionary for each type which contains a dictionary of all the JSON by
name.

Program is basically done now, all that's left is to write the screens.

Tags:[feat]