Refactored screen changing
Why:
Previously, I had to copy-paste the same screen into the program 6
times, change out the type names, and add them to the for loop at the
beginning of the program. This was rather redundant and not particularly
modular.

How:
Now, I changed the program so that mutations, items, monsters, etc. are
all one screen which changes out its values on the click of a button.

Tags:[ui][refactor]