Added a JSON translator
Why:
So that there is an easy way to translate JSON values into
human-readable values without hardcoding everything.

How:
For every attribute in a JSON object that is about to be outputted goes
to a special translation.json file and checks whether there is a
translation for that attribute. If there is, replaces the attribute with
the translation.

Tags:[feat][WIP]