Separated JSON searching into `JsonSearcher` class
Why:
It is far more modular and efficient like this. The searching didn't
really have any business in `LookupFrame` , and it has been taking up
an unreasonable amount of space in that class.

How:
Simply copied the functions for searching JSON in `LookupFrame` and
moved them to their own class.

Tags: