Added reading of JSON types from a file
Why:
So I don't have to hardcode what kind of type of entry a JSON entry is
(i.e. ARMOR => item) into the code, and so that the program is more
extensible.

How:
Added a new file which contains the json type=>program readable type
mapping, reads it, and turns it into a dictionary.

Tags:[feat]