Refactored JSON storage
Why:
So that it is easier to read JSON by attributes other than name, and so
that it is possible to read JSON attributes that don't have a name at
all.

How:
Rather than storing JSON in a dictionary of dictionaries, where the name
would act as the key to a specific entry, now stores JSON entries in a
dictionary of lists of dictionaries. This is a little less efficient,
but overall far more robust.

Tags:[refactor][fix]