Added further enhancements to recipe prettifier
Why:
As precedent for future similar work (which there will be large amounts
of)

How:
Created a new `organizedJson` attribute for JsonSearcher. It is the same
as `rawJson`, except all the entries are organized by their id. Then, in
the prettifier uses this attribute to look up the names of crafting
components (which are specified in json by id)

Tags:[feat]